[W-M]2013.7.30,蜂鸣器乱响，接收端也有错帧

+++ SYSTEM/usart/usart.h
@@ -40,7 +40,7 @@ void uart_init(u32 bound);
 #define EN_USART2_RX 			1		//使能（1）/禁止（0）串口1接收
 	  	
 extern u8  USART2_RX_BUF[USART2_REC_LEN]; //接收缓冲,最大USART_REC_LEN个字节.末字节为换行符 
-extern u8 USART2_RX_STA;         		//接收状态标记	
+extern u16 USART2_RX_STA;         		//接收状态标记	
 //如果想串口中断接收，请不要注释以下宏定义
 void uart2_init(u32 bound);
 #endif
